+++ src/main/resources/db/migration/V3__Artist.sql
@@ -1,6 +1,6 @@
 CREATE TABLE artists (
-    id BIGINT PRIMARY KEY AUTO_INCREMENT,
-    id_deezer BIGINT UNIQUE,
+    id INT PRIMARY KEY AUTO_INCREMENT,
+    id_deezer INT UNIQUE,
     name VARCHAR(255) UNIQUE NOT NULL,
     image VARCHAR(255),
     nb_album INT
